+++ lib_dominio/Entidades/Empleados.cs
@@ -6,7 +6,7 @@ namespace lib_dominio.Entidades
         public int Id { get; set; }
         public string? Nombre { get; set; }
         public string? Cedula { get; set; }
-        public double Salario { get; set; }
+        public decimal Salario { get; set; }
         public string? Telefono { get; set; }
     }
 }
